+++ CDEFramework/Libraries/NCILibrary/Code/NCILibrary.Search/ISiteWideSearchResultCollection.cs
@@ -7,42 +7,7 @@ namespace NCI.Search
 {
     public interface ISiteWideSearchResultCollection : IEnumerable<ISiteWideSearchResult>
     {
-        int ResultCount { get; set; }        
-
-        //ArrayList WordForms
-        //{
-        //    get;
-        //    set;
-        //}
-        //string DidYouMean
-        //{
-        //    get;
-        //    set;
-        //}
-        //long TotalNumResults
-        //{
-        //    get;
-        //    set;
-        //}
-
-        //void Add(ISiteWideSearchResult value);
-
-        //void Clear();
-
-        //bool Contains(ISiteWideSearchResult value);
-
-        //int IndexOf(ISiteWideSearchResult value);
-
-        //void Insert(int index, ISiteWideSearchResult value);
-
-        //void Remove(ISiteWideSearchResult value);
-
-        //void RemoveAt(int index);
-
-        //ISiteWideSearchResult this[int index]
-        //{
-        //    get;
-        //    set;
-        //}
+        long ResultCount { get; set; }        
+              
     }
 }
